Add `cap_eps` argument to `compute_metrics_at_error_threshold` which toggles whether to cap unreasonably large values of epsilon or not (#6)
Summary:
Pull Request resolved: https://github.com/facebookresearch/PrivacyGuard/pull/6

This diff introduces a new argument `cap_eps` to the `compute_metrics_at_error_threshold` method, which allows the user to toggle whether to cap unreasonably large values of epsilon or not. The default value of `cap_eps` is set to `True` to avoid impacting current prod readings. This makes it consistent with the method signature in `compute_eps_at_tpr_threshold`

Changes:

* Adding the `cap_eps` argument to the `compute_metrics_at_error_threshold` method in `mia_results.py`
* Passing the `cap_eps` argument from the `compute_metrics_at_error_threshold` method in `analysis_node.py` to the `compute_metrics_at_error_threshold` method in `mia_results.py`
* Updating the docstring in `mia_results.py` to reflect the new argument

Reviewed By: mgrange1998

Differential Revision: D77235199

fbshipit-source-id: 81f5966fc513f251c8e0f6db93ccce4c9f9839b2