Move analysis modules for MIA to subdirectory mia (#13)
Summary:
Pull Request resolved: https://github.com/facebookresearch/PrivacyGuard/pull/13

## Diff Summary

**Moving Analysis Modules for MIA to Subdirectory**

This diff is the second of three diffs that aim to move analysis modules for MIA to a subdirectory. The changes are as follows:

### Code Changes

#### `benchmark_factored_analysis_node.py`

*   Moved import statements to use `privacy_guard.analysis.mia` instead of `privacy_guard.analysis`.

#### `test_factored_analysis_node.py`

*   Updated import statement to use `privacy_guard.analysis.mia.factored_analysis_node` instead of `privacy_guard.analysis.factored_analysis_node`.

#### `compute_score.py`

*   Moved import statement to use `privacy_guard.analysis.mia.factored_analysis_node` instead of `privacy_guard.analysis.factored_analysis_node`.

#### `BUCK`

*   Updated the source file for the `factored_analysis_node` library to `mia/factored_analysis_node.py` instead of `factored_analysis_node.py`.

### Overall Impact

This diff is a part of a larger effort to restructure the analysis modules for MIA. By moving these modules to a subdirectory, the codebase becomes more organized and easier to maintain. The changes made in this diff are primarily related to updating import statements and build configurations to reflect the new directory structure.

Differential Revision: D80014423