LIA integration (#16)
Summary:
Pull Request resolved: https://github.com/facebookresearch/PrivacyGuard/pull/16

Add lightweight label inference attack to PrivacyGuard.

Note: I also define a LIAAttackInput class which prepares the datasets before feeding them into the attack. The purpose of this is to speed up experimentation. Aggregation for LIA is expensive (since it uses .idxmax(). Therefore, we want to avoid re-doing the aggregation every time we change something in the attack or analysis.

Reviewed By: lucamelis

Differential Revision: D77935914

fbshipit-source-id: 038958b3fbdf77f32bdee34051fedbe953e54b7a