Add custom user_id to AggregateAnalysisInput, CalibAttack, LiraAttack. "separable_id" in AggregateAnalysisInput
Summary:
This change adds a "user_id_key" argument to AggregateAnalysisInput, CalibAttack, and LiraAttack.

This will be used to sanitize references to "separable_id" in the OSS release, in subsequent diffs in this stack.
To keep the incremental changes simple, this change adds this new argument but keeps the default behavior of CalibAttack and LiraAttack to use "separable_id".

# Change Breakdown

- Adds new "user_id_key" to AggregateAnalysisInput, CalibAttack, and LiraAttack.
- Replaces hardcoded instances of "separable_id" with this variable
- Updates default behavior for CalibAttack and LiraAttack so that user_id_key is set to "separable_id" by default.

# Next steps

1. Migrate "separable_id" to "user_id" in OSS files, update testing datasets
2. Remove default altogether and update all callsites to specify "separable_id" explicitly.

Differential Revision: D80955348