Support for passing in tokenizer and model directly
Summary:
This is a refactor to GenerationAttack node that supports passing in tokenizer and model to the GenerationAttackCustomModel directly. This is to support cases where the models are loaded with something other than "AutoTokenizer" and "AutoModelForCausalLM", or pass in custom arguments like `cache_dir`

The GenerationAttack node is functionally equivalent to as it was before, but now inherits from GenerationAttackCustomModel and passes in the loaded tokenizer and model into the parent initializer.

Differential Revision: D82024621