Support for passing in tokenizer and model directly (#42)
Summary:
Pull Request resolved: https://github.com/facebookresearch/PrivacyGuard/pull/42

This is a refactor to GenerationAttack node that supports passing in tokenizer and model to the GenerationAttackCustomModel directly. This is to support cases where the models are loaded with something other than "AutoTokenizer" and "AutoModelForCausalLM", or pass in custom arguments like `cache_dir`

The GenerationAttack node is functionally equivalent to as it was before, but now inherits from GenerationAttackCustomModel and passes in the loaded tokenizer and model into the parent initializer.

Reviewed By: knchadha

Differential Revision: D82024621

fbshipit-source-id: 8104acfe668631bbe55f9ca2365f73e7cb036a32